+++ frontend/src/components/Logo.jsx
@@ -1,6 +1,6 @@
 const Logo = () => {
   return (
-    <div className="flex text-white/90 items-center">
+    <div className="flex text-white/90 items-center cursor-default">
       <span className="font-serif font-light text-4xl italic mr-3">{"{"}</span>
       <span className="font-serif font-medium text-2xl italic mr-2 text-violet-200">
         prompt: